pyproject.toml
[tool.poetry]
name = "ml-ollama"
version = "0.1.0"
description = "End-to-end ML project with Clean + Hexagonal Architecture"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9,<3.13"
# Data Processing
pandas = ">=2.0.0,<2.3.0"
numpy = ">=1.24.0,<1.27.0"
# ML & Science
scikit-learn = ">=1.3.0,<1.6.0"
scipy = ">=1.10.0,<1.14.0"
# PDF Processing
pypdf2 = ">=3.0.0"
pdfplumber = ">=0.10.0,<0.12.0"
# OCR for scanned PDFs
pytesseract = ">=0.3.10"
pillow = ">=10.0.0"
pdf2image = ">=1.16.3"
# Visualization
matplotlib = ">=3.7.0,<3.11.0"
seaborn = ">=0.12.0,<0.14.0"
plotly = ">=5.17.0"
# Utilities
pydantic = ">=2.4.0,<3.0.0"
python-dotenv = ">=1.0.0"
loguru = ">=0.7.0"
# CLI
typer = {extras = ["all"], version = "^0.20.0"}
rich = ">=13.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
isort = "^5.12.0"

[tool.poetry.scripts]
ml-pipeline = "src.presentation.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
